@@ -67,8 +67,8 @@
     "next-mdx-remote": "2.1.3",
     "next-themes": "0.0.12",
     "preact": "10.7.1",
-    "react": "18.1.0",
-    "react-dom": "18.1.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-icons": "4.3.1",
     "react-is": "17.0.1",
     "reading-time": "1.3.0",
@@ -85,21 +85,21 @@
     "@babel/preset-env": "7.13.9",
     "@babel/preset-react": "7.12.13",
     "@babel/preset-typescript": "7.13.0",
-    "@nrwl/cli": "14.1.1",
-    "@nrwl/cypress": "14.1.1",
-    "@nrwl/eslint-plugin-nx": "14.1.1",
-    "@nrwl/jest": "14.1.1",
-    "@nrwl/next": "14.1.1",
-    "@nrwl/nx-cloud": "14.0.3",
-    "@nrwl/react": "14.1.1",
-    "@nrwl/web": "14.1.1",
-    "@nrwl/workspace": "14.1.1",
-    "@testing-library/react": "13.1.1",
+    "@nrwl/cli": "14.4.3",
+    "@nrwl/cypress": "14.4.3",
+    "@nrwl/eslint-plugin-nx": "14.4.3",
+    "@nrwl/jest": "14.4.3",
+    "@nrwl/next": "14.4.3",
+    "@nrwl/nx-cloud": "14.2.0",
+    "@nrwl/react": "14.4.3",
+    "@nrwl/web": "14.4.3",
+    "@nrwl/workspace": "14.4.3",
+    "@testing-library/react": "13.3.0",
     "@types/jest": "27.4.1",
     "@types/mdx-js__react": "1.5.3",
-    "@types/node": "14.14.33",
-    "@types/react": "18.0.8",
-    "@types/react-dom": "18.0.3",
+    "@types/node": "18.0.0",
+    "@types/react": "18.0.14",
+    "@types/react-dom": "18.0.5",
     "@types/react-is": "17.0.3",
     "@types/react-mailchimp-subscribe": "2.1.1",
     "@typescript-eslint/eslint-plugin": "4.14.0",
@@ -116,31 +116,31 @@
     "eslint-config-prettier": "7.2.0",
     "eslint-plugin-cypress": "2.11.2",
     "eslint-plugin-import": "2.26.0",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.29.4",
-    "eslint-plugin-react-hooks": "4.5.0",
+    "eslint-plugin-jsx-a11y": "6.6.0",
+    "eslint-plugin-react": "7.30.1",
+    "eslint-plugin-react-hooks": "4.6.0",
     "globby": "11.0.2",
     "image-size": "0.9.4",
     "jest": "27.5.1",
     "npm-check-updates": "11.8.5",
-    "nx": "14.1.1",
+    "nx": "14.4.3",
     "postcss": "8.4.13",
     "postcss-flexbugs-fixes": "5.0.2",
     "postcss-import": "14.1.0",
     "postcss-preset-env": "7.5.0",
-    "prettier": "2.6.2",
-    "react-is": "18.1.0",
-    "react-test-renderer": "18.1.0",
+    "prettier": "2.7.1",
+    "react-is": "18.2.0",
+    "react-test-renderer": "18.2.0",
     "rehype": "11.0.0",
     "remark-frontmatter": "3.0.0",
     "remark-parse": "9.0.0",
     "remark-stringify": "9.0.1",
     "rss": "1.2.2",
     "ts-jest": "27.1.4",
-    "ts-node": "9.1.1",
+    "ts-node": "10.8.2",
     "tsconfig-paths": "3.9.0",
     "tslint": "6.1.3",
-    "typescript": "4.4.4",
+    "typescript": "4.7.4",
     "unified": "9.2.1",
     "unist-util-visit": "2.0.3"
   }
