@@ -1,6 +1,6 @@
 'use strict';
 
-module.exports = {
+export default {
   displayName: 'website',
 
   transform: {
@@ -9,5 +9,5 @@ module.exports = {
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
   coverageDirectory: '../../coverage/apps/website',
-  preset: '../../jest.preset.ts',
+  preset: '../../jest.preset.js',
 };
